Blacklist optimisation (#549)
Adding the isBlacklisted flag to the IndexedNFT model to optimise query performance.
The NFT Blacklist indexer will still operate independently but occasionally the flags will be synced from there into the NFTS collection.
N.B. The isBlacklisted flag on IndexedNFT is treated as a metadata field and is not versioned.

Removing thor-solo as part of the default configuration for dev. Most devs will use a mainnet or testnet node. Thor solo is only required for integration and e2e tests so I am treating it as such.

Bug fix on parseRecords for NFTs. Duplicate records were not being handled correctly. (It looks like there is limited or no impact because of this in production.)

Fixing the pruner. It wasn't triggering for log indexers